refactor: extract module graph serialization
